requirements.txt
pandas
sqlalchemy
psycopg2-binary
schedule
plotly
python-dotenv
numpy
loguru